fix locale migration and migrate ssh host key

+++ recipes-calaos/calaos/calaos-server/calaos_migrate.sh
@@ -33,30 +33,30 @@ echo "[*] Searching for an old calaos installation..."
 cdrive=""
 while IFS='' read -r line || [[ -n "$line" ]]
 do
-	drive=$(echo $line | cut -d' ' -f4)
-	if [ "${drive:0:3}" == "ram" ]; then continue; fi
-	if [ "${drive:0:4}" == "loop" ]; then continue; fi
-	
-	if [ ! -b /dev/$drive ]; then continue; fi
-	
-	set +e
-	mount /dev/$drive $tmpdir > /dev/null 2>&1
-	if [ $? -eq 0 ]; then
-		if [ -d $tmpdir/etc/calaos ]
-		then
-			cdrive=$drive
-			break
-		fi
-	fi
-	set -e
+    drive=$(echo $line | cut -d' ' -f4)
+    if [ "${drive:0:3}" == "ram" ]; then continue; fi
+    if [ "${drive:0:4}" == "loop" ]; then continue; fi
+    
+    if [ ! -b /dev/$drive ]; then continue; fi
+    
+    set +e
+    mount /dev/$drive $tmpdir > /dev/null 2>&1
+    if [ $? -eq 0 ]; then
+        if [ -d $tmpdir/etc/calaos ]
+        then
+            cdrive=$drive
+            break
+        fi
+    fi
+    set -e
 done < /proc/partitions
 
 if [ x$cdrive == x ]
 then
-	echo "No calaos installation found! Cannot migrate."
-	echo "Are you sure a calaos-os was installed?"
-	echo "Aborting."
-	exit 1
+    echo "No calaos installation found! Cannot migrate."
+    echo "Are you sure a calaos-os was installed?"
+    echo "Aborting."
+    exit 1
 fi
 
 echo
@@ -91,19 +91,19 @@ dir=$(pwd)
 cd $tmpdir/etc/systemd/system/multi-user.target.wants/
 
 is_service() {
-	if [ "$1" == "calaos-home.service" ]; then return 0; fi
-	if [ "$1" == "calaos-server.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyUSB0.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyUSB1.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyUSB2.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyUSB3.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyS0.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyS1.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyS2.service" ]; then return 0; fi
-	if [ "$1" == "usb-serial-touchscreen@ttyS3.service" ]; then return 0; fi
-	if [ "$1" == "olad.service" ]; then return 0; fi
-	if [ "$1" == "knxd.service" ]; then return 0; fi
-	return 1
+    if [ "$1" == "calaos-home.service" ]; then return 0; fi
+    if [ "$1" == "calaos-server.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyUSB0.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyUSB1.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyUSB2.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyUSB3.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyS0.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyS1.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyS2.service" ]; then return 0; fi
+    if [ "$1" == "usb-serial-touchscreen@ttyS3.service" ]; then return 0; fi
+    if [ "$1" == "olad.service" ]; then return 0; fi
+    if [ "$1" == "knxd.service" ]; then return 0; fi
+    return 1
 }
 
 for s in *.service
@@ -136,10 +136,10 @@ echo "[*] Copy touchscreen calibration data"
 dir=$(pwd)
 
 is_xconf() {
-	if [ "$1" == "10-evdev.conf" ]; then return 0; fi
-	if [ "$1" == "10-quirks.conf" ]; then return 0; fi
-	if [ "$1" == "50-synaptics.conf" ]; then return 0; fi
-	return 1
+    if [ "$1" == "10-evdev.conf" ]; then return 0; fi
+    if [ "$1" == "10-quirks.conf" ]; then return 0; fi
+    if [ "$1" == "50-synaptics.conf" ]; then return 0; fi
+    return 1
 }
 
 cd $tmpdir/usr/share/X11/xorg.conf.d
@@ -150,8 +150,8 @@ do
         continue
     fi
 
-	echo "  copying $c"
-	test -d /usr/share/X11/xorg.conf.d && cp $tmpdir/usr/share/X11/xorg.conf.d/$c /usr/share/X11/xorg.conf.d/$c
+    echo "  copying $c"
+    test -d /usr/share/X11/xorg.conf.d && cp $tmpdir/usr/share/X11/xorg.conf.d/$c /usr/share/X11/xorg.conf.d/$c
 done
 cd $dir
 
@@ -159,21 +159,25 @@ echo "[*] Copy ssh keys and root password"
 test -d $tmpdir/home/root/.ssh && cp -R $tmpdir/home/root/.ssh /home/root/
 rootpw=$(cat $tmpdir/etc/shadow | grep "root:")
 sed -i "s/^root:.*\$/$(echo $rootpw | sed -e 's/[]\/$*.^|[]/\\&/g')/g" /etc/shadow
+test -d $tmpdir/etc/dropbear && cp -R $tmpdir/etc/dropbear /etc/
 
 echo "[*] Setup timezone"
 cp -P $tmpdir/etc/localtime /etc/localtime
 systemctl stop ntpd
 systemctl start ntpd
 
-echo "[*] Setup locale"
-cp $tmpdir/etc/locale.conf /etc/locale.conf
-lc=$(cat $tmpdir/etc/locale.conf | cut -d= -f2)
-km=$(cat $tmpdir/etc/vconsole.conf | cut -d= -f2)
-xk=$(cat $tmpdir/etc/X11/xorg.conf.d/00-keyboard.conf | grep XkbLayout | cut -d\" -f4)
+if [ -e $tmpdir/etc/locale.conf ]
+then
+    echo "[*] Setup locale"
+    cp $tmpdir/etc/locale.conf /etc/locale.conf
+    lc=$(cat $tmpdir/etc/locale.conf | cut -d= -f2)
+    km=$(cat $tmpdir/etc/vconsole.conf | cut -d= -f2)
+    xk=$(cat $tmpdir/etc/X11/xorg.conf.d/00-keyboard.conf | grep XkbLayout | cut -d\" -f4)
 
-localectl set-locale LANG="$lc"
-localectl set-keymap $km
-localectl set-x11-keymap $xk
+    localectl set-locale LANG="$lc"
+    localectl set-keymap $km
+    localectl set-x11-keymap $xk
+fi
 
 echo "[*] Install SSL certificate"
 if [ -e $tmpdir/etc/ssl/haproxy/server.pem ]
@@ -186,14 +190,14 @@ fi
 
 if [ -e $tmpdir/etc/default/eibnetmux ]
 then
-	echo "[*] Copy eibnetmux configuration"
-	cp $tmpdir/etc/default/eibnetmux /etc/default/eibnetmux
+    echo "[*] Copy eibnetmux configuration"
+    cp $tmpdir/etc/default/eibnetmux /etc/default/eibnetmux
 fi
 
 if [ -d $tmpdir/etc/ola ]
 then
-	echo "[*] Copy OLA configuration"
-	cp -R $tmpdir/etc/ola /etc/
+    echo "[*] Copy OLA configuration"
+    cp -R $tmpdir/etc/ola /etc/
 fi
 
 echo "[*] Copy network configuration"
